Feature/pilhwan

+++ backend/src/main/java/OSSP/demo/service/notice/NoticeService.java
@@ -19,6 +19,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @Slf4j
@@ -33,32 +34,43 @@ public class NoticeService {
     //각 메서드마다 올바르지 않은 사용자(팀이 아닌 사람이 생성한)에 대해 예외처리
     @Transactional(readOnly = true)
     public ResponseEntity findAll(String studentId, Long teamId) {
-
-        User user = userRepository.findByStudentId(studentId).get();
         try {
-            Member member = memberRepository.findByUserIdAndTeamId(user.getId(), teamId).get();
-            List<Notice> notices = noticeRepository.findByTeamId(teamId);
-
-            List<NoticeResponse> responseDto = new ArrayList<>();
-            for (Notice notice : notices) {
-                NoticeResponse dto = NoticeResponse.builder()
-                        .id(notice.getNoticeId())
-                        .content(notice.getContent())
-                        .build();
-                responseDto.add(dto);
+            User user = userRepository.findByStudentId(studentId).get();
+            Member member = memberRepository.findByUserIdAndTeamId(user.getId(), teamId).orElse(null);
+            if (member==null){
+                return ResponseEntity.badRequest().body(ResponseDto.builder().error(Collections.singletonMap("get_notice", "올바른 사용자가 아닙니다.")).build());
             }
-            return ResponseEntity.ok().body(responseDto);
+            List<Notice> notices = noticeRepository.findByTeamId(teamId);
+            List<NoticeResponse> responseDto = notices.stream()
+                    .map(n -> new NoticeResponse(n))
+                    .collect(Collectors.toList());
+//            for (Notice notice : notices) {
+//                NoticeResponse dto = NoticeResponse.builder()
+//                        .id(notice.getNoticeId())
+//                        .content(notice.getContent())
+//                        .build();
+//                responseDto.add(dto);
+//            }
+            return ResponseEntity.ok().body(Collections.singletonMap("get_notices", responseDto));
         } catch (Exception e) {
             log.error(e.getMessage());
-            return ResponseEntity.badRequest().body(ResponseDto.builder().error(Collections.singletonMap("write_notice", "공지사항 조회에 실패했습니다.(올바르지 않은 사용자)")).build());
+            ResponseDto responseErrorDto = ResponseDto.builder().error(Collections.singletonMap("get_notice", "공지사항 조회에 실패했습니다.")).build();
+            return ResponseEntity.badRequest().body(responseErrorDto);
         }
     }
 
     @Transactional
     public ResponseEntity write(String studentId, Long teamId, NoticeRequest noticeRequest){
-        User user = userRepository.findByStudentId(studentId).get();
         try{
-            Member member = memberRepository.findByUserIdAndTeamId(user.getId(), teamId).get();
+            User user = userRepository.findByStudentId(studentId).get();
+            Member member = memberRepository.findByUserIdAndTeamId(user.getId(), teamId).orElse(null);
+            if (member==null){
+                return ResponseEntity.badRequest().body(ResponseDto.builder().error(Collections.singletonMap("write_notice", "올바른 사용자가 아닙니다.")).build());
+            }
+            String content = noticeRequest.getContent();
+            if (content==null){
+                return ResponseEntity.badRequest().body(ResponseDto.builder().error(Collections.singletonMap("write_notice", "내용을 입력하시오")).build());
+            }
             Notice notice = noticeRequest.toEntity();
             notice.setMember(member);
             noticeRepository.save(notice);
@@ -66,20 +78,26 @@ public class NoticeService {
                     .id(notice.getNoticeId())
                     .content(notice.getContent())
                     .build();
-            return ResponseEntity.ok().body(Collections.singletonMap("write_notice", "공지사항을 생성했습니다."));
+            return ResponseEntity.ok().body(dto);
         }catch (Exception e){
             log.error(e.getMessage());
-            return ResponseEntity.badRequest().body(ResponseDto.builder().error(Collections.singletonMap("write_notice", "공지사항 생성에 실패했습니다.(올바르지 않은 사용자)")).build());
-
+            ResponseDto responseErrorDto = ResponseDto.builder().error(Collections.singletonMap("write_notice", "공지사항 생성에 실패했습니다.")).build();
+            return ResponseEntity.badRequest().body(responseErrorDto);
         }
-
     }
 
     @Transactional
     public ResponseEntity update(String studentId, Long teamId, Long noticeId, NoticeRequest noticeRequest) {
-        User user = userRepository.findByStudentId(studentId).get();
         try {
-            Member member = memberRepository.findByUserIdAndTeamId(user.getId(), teamId).get();
+            User user = userRepository.findByStudentId(studentId).get();
+            Member member = memberRepository.findByUserIdAndTeamId(user.getId(), teamId).orElse(null);
+            if (member==null){
+                return ResponseEntity.badRequest().body(ResponseDto.builder().error(Collections.singletonMap("update_notice", "올바른 사용자가 아닙니다.")).build());
+            }
+            String content = noticeRequest.getContent();
+            if (content==null){
+                return ResponseEntity.badRequest().body(ResponseDto.builder().error(Collections.singletonMap("update_notice", "수정할 내용을 입력하시오")).build());
+            }
             Notice findNotice = noticeRepository.findById(noticeId).get();
             Notice notice = noticeRequest.toEntity();
             findNotice.updateContent(notice.getContent(), member);
@@ -89,27 +107,29 @@ public class NoticeService {
                     .content(findNotice.getContent())
                     .build();
 
-            return ResponseEntity.ok().body(Collections.singletonMap("update_notice", "공지사항을 수정했습니다."));
+            return ResponseEntity.ok().body(dto);
         } catch (Exception e) {
             log.error(e.getMessage());
-            return ResponseEntity.badRequest().body(ResponseDto.builder().error(Collections.singletonMap("update_notice", "공지사항 수정에 실패했습니다.(올바르지 않은 사용자)")).build());
+            ResponseDto responseErrorDto = ResponseDto.builder().error(Collections.singletonMap("update_notice", "공지사항 수정에 실패했습니다.")).build();
+            return ResponseEntity.badRequest().body(responseErrorDto);
         }
     }
 
     @Transactional
     public ResponseEntity delete(String studentId, Long teamId, Long noticeId){
-        User user = userRepository.findByStudentId(studentId).get();
         try {
-            Member member = memberRepository.findByUserIdAndTeamId(user.getId(), teamId).get();
+            User user = userRepository.findByStudentId(studentId).get();
+            Member member = memberRepository.findByUserIdAndTeamId(user.getId(), teamId).orElse(null);
+
+            if (member==null){
+                return ResponseEntity.badRequest().body(ResponseDto.builder().error(Collections.singletonMap("delete_notice", "올바른 사용자가 아닙니다.")).build());
+            }
             noticeRepository.deleteById(noticeId);
             return ResponseEntity.ok().body(Collections.singletonMap("delete_notice", "공지사항 삭제에 성공했습니다."));
         }catch (Exception e){
             log.error(e.getMessage());
-            return ResponseEntity.badRequest().body(ResponseDto.builder().error(Collections.singletonMap("delete_notice", "삭제를 실패했습니다.")).build());
+            ResponseDto responseErrorDto = ResponseDto.builder().error(Collections.singletonMap("delete_notice", "공지사항 삭제에 실패했습니다.")).build();
+            return ResponseEntity.badRequest().body(responseErrorDto);
         }
-
     }
-
-
-
 }
